personal_portfolio/requirements.txt
certifi==2019.9.11
Django==2.2.6
django-tastypie==0.14.2
djangorestframework==3.10.3
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
wincertstore==0.2
